[CHORE] data class 분리 및 InviteCategory 생성
- data class 분리(Category, TodoData)
- data class InviteCategory 생성

Fixes: TICO-31
