+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/data/local/datasource/DataSource.kt
@@ -1,8 +1,9 @@
 package com.tico.pomorodo.data.local.datasource
 
+import com.tico.pomorodo.data.local.entity.Category
+import com.tico.pomorodo.data.local.entity.InviteCategory
+import com.tico.pomorodo.data.local.entity.TodoData
 import com.tico.pomorodo.data.local.entity.User
-import com.tico.pomorodo.ui.todo.view.Category
-import com.tico.pomorodo.ui.todo.view.TodoData
 import com.tico.pomorodo.ui.todo.view.TodoState
 
 object DataSource {
@@ -24,28 +25,33 @@ object DataSource {
 
     val todoList = listOf<Category>(
         Category(
+            id = "1",
             title = "카테고리 1",
             groupNumber = 5,
             todoList = listOf(
                 TodoData(
+                    id = "1",
                     name = "Todo 1",
                     state = TodoState.UNCHECKED,
                     completeGroupNumber = 4,
                     likedNumber = 22
                 ),
                 TodoData(
+                    id = "2",
                     name = "Todo 1",
                     state = TodoState.CHECKED,
                     completeGroupNumber = 1,
                     likedNumber = 22
                 ),
                 TodoData(
+                    id = "3",
                     name = "Todo 1",
                     state = TodoState.GOING,
                     completeGroupNumber = 3,
                     likedNumber = 22
                 ),
                 TodoData(
+                    id = "4",
                     name = "Todo 1",
                     state = TodoState.UNCHECKED,
                     completeGroupNumber = 0,
@@ -54,25 +60,31 @@ object DataSource {
             )
         ),
         Category(
+            id = "2",
             title = "카테고리 2",
             todoList = listOf(
                 TodoData(
+                    id = "1",
                     name = "Todo 1",
                     state = TodoState.UNCHECKED,
                     completeGroupNumber = 4,
                     likedNumber = 22
                 ),
             )
-        ), Category(
+        ),
+        Category(
+            id = "3",
             title = "카테고리 3",
             todoList = listOf(
                 TodoData(
+                    id = "1",
                     name = "Todo 1",
                     state = TodoState.UNCHECKED,
                     completeGroupNumber = 4,
                     likedNumber = 22
                 ),
                 TodoData(
+                    id = "2",
                     name = "Todo 1",
                     state = TodoState.CHECKED,
                     completeGroupNumber = 4,
@@ -80,5 +92,18 @@ object DataSource {
                 ),
             )
         )
+    )
+
+    val inviteList = listOf<InviteCategory>(
+        InviteCategory(id = "1", title = "알고리즘 뿌셔!", groupReader ="랑드샤 쿠키"),
+        InviteCategory(id = "2", title = "공작 영애들의 우아한 코딩 궁전", groupReader ="모카커피1"),
+        InviteCategory(id = "3", title = "공작 영애들의 우아한 코딩 궁전", groupReader ="모카커피2"),
+        InviteCategory(id = "4", title = "공작 영애들의 우아한 코딩 궁전", groupReader ="모카커피3"),
+        InviteCategory(id = "5", title = "알고리즘 뿌셔!", groupReader ="모카커피4"),
+        InviteCategory(id = "6", title = "공작 영애들의 우아한 코딩 궁전", groupReader ="모카커피1"),
+        InviteCategory(id = "7", title = "공작 영애들의 우아한 코딩 궁전", groupReader ="모카커피1"),
+        InviteCategory(id = "7", title = "공작 영애들의 우아한 코딩 궁전", groupReader ="모카커피1"),
+        InviteCategory(id = "7", title = "공작 영애들의 우아한 코딩 궁전", groupReader ="모카커피1"),
+        InviteCategory(id = "7", title = "공작 영애들의 우아한 코딩 궁전", groupReader ="모카커피1"),
     )
 }
