+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/data/local/entity/Category.kt
@@ -1,0 +1,8 @@
+package com.tico.pomorodo.data.local.entity
+
+data class Category(
+    val id: String,
+    val title: String,
+    val todoList: List<TodoData>,
+    val groupNumber: Int = 0
+)
