+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/data/local/entity/InviteCategory.kt
@@ -1,0 +1,7 @@
+package com.tico.pomorodo.data.local.entity
+
+data class InviteCategory(
+    val id: String,
+    val title: String,
+    val groupReader: String
+)
