+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/data/local/entity/TodoData.kt
@@ -1,0 +1,11 @@
+package com.tico.pomorodo.data.local.entity
+
+import com.tico.pomorodo.ui.todo.view.TodoState
+
+data class TodoData(
+    val id: String,
+    val name: String,
+    var state: TodoState,
+    val completeGroupNumber: Int? = null,
+    val likedNumber: Int = 0,
+)
