+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/ui/todo/view/TodoScreen.kt
@@ -34,25 +34,13 @@ import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.compose.rememberNavController
 import com.tico.pomorodo.data.local.datasource.DataSource
+import com.tico.pomorodo.data.local.entity.Category
 import com.tico.pomorodo.data.local.entity.User
 import com.tico.pomorodo.ui.AppState
 import com.tico.pomorodo.ui.common.view.BottomBar
 import com.tico.pomorodo.ui.common.view.addFocusCleaner
 import com.tico.pomorodo.ui.theme.PomoroDoTheme
 import com.tico.pomorodo.ui.todo.viewmodel.TodoViewModel
-
-data class TodoData(
-    val name: String,
-    var state: TodoState,
-    val completeGroupNumber: Int? = null,
-    val likedNumber: Int = 0,
-)
-
-data class Category(
-    val title: String,
-    val todoList: List<TodoData>,
-    val groupNumber: Int = 0
-)
 
 @Composable
 fun TodoScreen(
